requirements.txt
Django==2.0.6
djangorestframework==3.8.2
geographiclib==1.49
geopy==1.14.0
polyline==1.3.2
pytz==2018.4
six==1.11.0
